:recycle: [Refector] 코드 분리작업
- Component 분리
- index 추가
