+++ index.html
@@ -34,6 +34,6 @@
         </div>
       </main>
     </div>
-    <script type="module" src="./src/js/App.js"></script>
+    <script type="module" src="./src/js/index.js"></script>
   </body>
 </html>
