+++ src/js/enum.js
@@ -1,0 +1,1 @@
+export const KEYCODE_ENTER = 13;
