+++ src/js/index.js
@@ -1,0 +1,3 @@
+import TodoApp from './Components/TodoApp.js';
+
+new TodoApp();
